refactor typescript code in separate modules
