+++ src/widget_core.ts
@@ -1,0 +1,41 @@
+import * as tone from 'tone';
+
+import { AudioNodeModel } from './widget_base';
+
+export class DestinationModel extends AudioNodeModel {
+  defaults(): any {
+    return {
+      ...super.defaults(),
+      _model_name: DestinationModel.model_name,
+      mute: false,
+      volume: -16,
+    };
+  }
+
+  createNode(): tone.ToneAudioNode {
+    return tone.getDestination();
+  }
+
+  get mute(): boolean {
+    return this.get('mute');
+  }
+
+  get volume(): number {
+    return this.get('volume');
+  }
+
+  initEventListeners(): void {
+    super.initEventListeners();
+
+    this.on('change:mute', () => {
+      this.node.mute = this.mute;
+    });
+    this.on('change:volume', () => {
+      this.node.volume.value = this.volume;
+    });
+  }
+
+  node: typeof tone.Destination;
+
+  static model_name = 'DestinationModel';
+}
